+++ Rdmp.Core/Providers/Nodes/FrozenExtractionConfigurationsNode.cs
@@ -4,6 +4,8 @@
 // RDMP is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
+using System;
+using Rdmp.Core.Curation.Data.Cohort;
 using Rdmp.Core.DataExport.Data;
 
 namespace Rdmp.Core.Providers.Nodes
@@ -11,7 +13,7 @@ namespace Rdmp.Core.Providers.Nodes
     /// <summary>
     /// Collection of all previously extracted (and now readonly) <see cref="ExtractionConfiguration"/>s in a given <see cref="Project"/>
     /// </summary>
-    class FrozenExtractionConfigurationsNode:Node
+    class FrozenExtractionConfigurationsNode:Node , IOrderable
     {
         public Project Project { get; set; }
 
@@ -42,5 +44,7 @@ namespace Rdmp.Core.Providers.Nodes
         {
             return (Project != null ? Project.GetHashCode() : 0);
         }
+
+        public int Order { get { return 6000; } set{} }
     }
 }