+++ Tools/rdmp/Program.cs
@@ -24,6 +24,7 @@ using Rdmp.Core.Startup;
 using ReusableLibraryCode;
 using ReusableLibraryCode.Checks;
 using YamlDotNet.RepresentationModel;
+using YamlDotNet.Serialization;
 
 namespace Rdmp.Core
 {
@@ -97,7 +98,7 @@ namespace Rdmp.Core
                             (CohortCreationOptions opts) => Run(opts),
                             (PackOptions opts) => Run(opts),
                             (PlatformDatabaseCreationOptions opts) => Run(opts),
-                            (ExecuteCommandOptions opts) => Run(opts),
+                            (ExecuteCommandOptions opts) => RunCmd(opts),
                             (ConsoleGuiOptions opts) => Run(opts),
                             errs => 1);
 
@@ -132,6 +133,38 @@ namespace Rdmp.Core
             return 0;
         }
 
+        private static int RunCmd(ExecuteCommandOptions opts)
+        {
+            if(!string.IsNullOrWhiteSpace(opts.File))
+            {
+                if(!File.Exists(opts.File))
+                {
+                    Console.WriteLine($"Could not find file '{opts.File}'");
+                    return -55;
+                }
+
+                var content = File.ReadAllText(opts.File);
+
+                if (string.IsNullOrWhiteSpace(content))
+                {
+                    Console.WriteLine($"File is empty ('{opts.File}')");
+                    return -56;
+                }
+
+                try
+                {
+                    var d = new Deserializer();
+                    opts.Script = d.Deserialize<RdmpScript>(content);
+                }
+                catch(Exception ex)
+                {
+                    Console.WriteLine($"Error deserializing '{opts.File}': {ex.Message}");
+                    return -57;                    
+                }
+            }
+
+            return Run((RDMPCommandLineOptions) opts);
+        }
 
         private static int Run(RDMPCommandLineOptions opts)
         {