+++ CHANGELOG.md
@@ -8,6 +8,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+## [4.1.3] - 2020-06-15
+
+### Added
+
+- Added `-f` option to CLI (`rdmp.exe -f somefile.yaml`) to run all commands in a file
+- Added "Go To" to tab right click context menu (previously only available in collections).
+- Private key encryption file location can now be customized per user by setting an environment variable `RDMP_KEY_LOCATION`.  This will override any key file location specified in the RDMP platform database.
+
+### Changed
+
+- Frozen Extraction Configurations folder always appears at the bottom of the branch under Projects
+- Improved layout of query building errors in QueryBuilder SQL viewing user interfaces
+
+### Fixed
+
+- Fixed bug in tree ordering when comparing a fixed order node to a non fixed order node.
+
 ## [4.1.2] - 2020-06-03
 
 ### Added
@@ -422,7 +439,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed Culture (e.g. en-us) not being passed correctly in DelimitedFlatFileAttacher
 - Fixed bug where Updater would show older versions of RDMP as installable 'updates'
 
-[Unreleased]: https://github.com/HicServices/RDMP/compare/v4.1.2...develop
+[Unreleased]: https://github.com/HicServices/RDMP/compare/v4.1.3...develop
+[4.1.3]: https://github.com/HicServices/RDMP/compare/v4.1.2...v4.1.3
 [4.1.2]: https://github.com/HicServices/RDMP/compare/v4.1.1...v4.1.2
 [4.1.1]: https://github.com/HicServices/RDMP/compare/v4.1.0...v4.1.1
 [4.1.0]: https://github.com/HicServices/RDMP/compare/v4.0.3...v4.1.0